+++ examples/attribute_server/example_attributes_client_with_auth.py
@@ -16,7 +16,7 @@ from labfreed.pac_attributes.api_data_models.response import Attribute, TextAttr
 from labfreed.pac_attributes.pythonic.attribute_server_factory import AttributeServerFactory, Webframework
 from labfreed.pac_attributes.server.attribute_data_sources import Dict_DataSource
 from labfreed.pac_attributes.server.translation_data_sources import DictTranslationDataSource
-from labfreed.pac_attributes.well_knonw_attribute_keys import MetaAttributeKeys
+from labfreed.pac_attributes.well_known_attribute_keys import MetaAttributeKeys
 from labfreed.utilities.translations import Terms, Term
 
 from labfreed.pac_attributes.client.auth import AuthRule, static_credential