+++ labfreed/pac_attributes/pythonic/attribute_server_factory.py
@@ -1,5 +1,4 @@
 from enum import Enum
-from functools import wraps
 import json
 import logging
 from typing import Any, Protocol